Bots change color
Now the bots change color to red when they cannot move because of an obstacle (and back to green when they can). The idea is that we will change the color of the 2 horizontal lines on the model only, but currently, we have only one material. I am not sure how to set more than one material on a model, but if you do, you can retrieve the material through its index in the table.