Exit With numbers
Exit now shows the number of bots that must still pass to finish the game. Please set the scale for X and Y to -1 in the prefab (to fix the upside down problem) and set the textures array to size 9. Each element is (in order) a number sprite from the level select.